qtdemux: Update variables before early exit
This is an update of d78d5896272d78df41e696fac929e7dfb3bb3dfa

We still exit as early as possible in case of non-ok/non-unlinked combined
flow, but we first make sure that we update the internal position variables.

This ensures that if upstreams "ignores" the flow return (and carries on pushing),
we don't end up processing data with completely bogus variables/positions.