completion: Remove incorrect fixmes
In fact the fixmes are incorrect since these are options that are
available when built against gst-validate, which I didn't. There is
reference to these options in the HELP_SUMMARY that refers to these
options stating "if ges-launch is built with gst-validate ..." and these
get picked by the regex that list the options.