qtmux: Fix timescale of timecode tracks
They should have ideally the same timescale of the video track, which we
can't guarantee here as in theory timecode configuration and video
framerate could be different. However we should set a correct timescale
based on the framerate given in the timecode configuration, and not just
use the framerate numerator.

Different than the 1.12 version of the patch, this uses the previous,
suboptimal calculation of the timescale for video tracks to be
consistent with the video track.